+++ upload/installer/sql/2.1.7.sql
@@ -1,3 +1,0 @@
-ALTER TABLE {IFW_TBL_COMMENTS} ADD guest_name VARCHAR( 100 ) NOT NULL default '' ,
-ADD guest_email VARCHAR( 100 ) NOT NULL default '' ,
-ADD guest_url VARCHAR( 150 ) NOT NULL default '' ;